visibility changes and small improvements

+++ ORM_Lib/TypeMapper/PostgresTypeMapper.cs
@@ -46,7 +46,7 @@ namespace ORM_Lib.TypeMapper
             TypeMap.TryGetValue(type, out var a);
             return a;
         }
-        public string GetForeignKeyType() => "integer";
+        public string GetForeignKeyType() => "bigint";
 
         public string GetEnumType() => "text";
